move change-kdf into KM ownership
